--- src/main/java/com/readme/novels/dto/NovelsSearchParamDto.java
@@ -1,20 +1,0 @@
-package com.readme.novels.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.ToString;
-
-@Getter
-@Builder
-@ToString
-@AllArgsConstructor
-@NoArgsConstructor
-public class NovelsSearchParamDto {
-
-    private String author;
-    private String title;
-    private Integer page;
-
-}
