+++ src/main/java/com/readme/novels/episodes/model/Episodes.java
@@ -1,0 +1,33 @@
+package com.readme.novels.episodes.model;
+
+import java.time.LocalDateTime;
+import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Entity
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class Episodes {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+    private Long novelsId;
+    private String title;
+    private String content;
+    private Long views;
+    private Boolean free;
+    private LocalDateTime registration;
+    private LocalDateTime createDate;
+    private LocalDateTime updateDate;
+    private String status;
+}
