+++ src/main/java/com/readme/novels/episodes/repository/EpisodesRepository.java
@@ -1,0 +1,8 @@
+package com.readme.novels.episodes.repository;
+
+import com.readme.novels.episodes.model.Episodes;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface EpisodesRepository extends JpaRepository<Episodes, Long> {
+
+}
