+++ src/main/java/com/readme/novels/episodes/service/EpisodesService.java
@@ -1,0 +1,7 @@
+package com.readme.novels.episodes.service;
+
+import com.readme.novels.episodes.dto.EpisodesDto;
+
+public interface EpisodesService {
+    void addEpisodes(EpisodesDto episodesDto);
+}
