+++ src/main/java/com/readme/novels/novels/aws/AwsS3Service.java
@@ -1,0 +1,63 @@
+package com.readme.novels.novels.aws;
+
+import com.amazonaws.services.s3.AmazonS3;
+import com.amazonaws.services.s3.model.CannedAccessControlList;
+import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.amazonaws.services.s3.model.PutObjectRequest;
+import java.io.IOException;
+import java.io.InputStream;
+import java.util.UUID;
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpStatus;
+import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import org.springframework.web.server.ResponseStatusException;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class AwsS3Service {
+
+    @Value("${cloud.aws.s3.bucket}")
+    private String bucket;
+    @Value("${s3Path}")
+    private String s3Path;
+
+    private final AmazonS3 amazonS3;
+
+    public String uploadFile(MultipartFile multipartFile) {
+        String fileName = createFileName(multipartFile.getOriginalFilename());
+        ObjectMetadata objectMetadata = new ObjectMetadata();
+        objectMetadata.setContentLength(multipartFile.getSize());
+        objectMetadata.setContentType(multipartFile.getContentType());
+
+        try (InputStream inputStream = multipartFile.getInputStream()) {
+            amazonS3.putObject(new PutObjectRequest(bucket, fileName, inputStream, objectMetadata)
+                .withCannedAcl(CannedAccessControlList.PublicRead));
+        } catch (IOException e) {
+            throw new ResponseStatusException(HttpStatus.INTERNAL_SERVER_ERROR, "파일 업로드에 실패 했습니다");
+        }
+
+        String returnFileName = s3Path+fileName;
+
+        log.info("@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@@ {}", returnFileName);
+
+        return returnFileName;
+    }
+
+    // 파일명 난수화 (이전에 올린 이미지 또 올릴수도)
+    private String createFileName(String filename) {
+        return UUID.randomUUID().toString().concat(getFileExtension(filename));
+    }
+
+    // 파일 형식 .뒤에 삭제
+    private String getFileExtension(String fileName) {
+        try {
+            return fileName.substring(fileName.lastIndexOf("."));
+        } catch (StringIndexOutOfBoundsException e) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "잘못된 형식의 파일 입니다.");
+        }
+    }
+}
