+++ src/main/java/com/readme/novels/novels/dto/PaginationDto.java
@@ -1,0 +1,20 @@
+package com.readme.novels.novels.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class PaginationDto {
+    private int page;
+    private int size;
+    private long totalElements;
+    private int totalPage;
+
+}
