+++ src/main/java/com/readme/novels/novels/responseObject/Response.java
@@ -1,0 +1,14 @@
+package com.readme.novels.novels.responseObject;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class Response {
+
+    private Object data;
+
+}
