+++ src/main/java/com/readme/novels/novels/responseObject/ResponseNovelsPagination.java
@@ -1,0 +1,28 @@
+package com.readme.novels.novels.responseObject;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.ToString;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+@ToString
+public class ResponseNovelsPagination {
+    Object contents;
+    Object pagination;
+
+    @Getter
+    @Builder
+    @NoArgsConstructor
+    @AllArgsConstructor
+    public static class Pagination {
+        private int page;
+        private int size;
+        private long totalElements;
+        private int totalPage;
+    }
+}
