--- src/main/java/com/readme/novels/repository/INovelsRepository.java
@@ -1,8 +1,0 @@
-package com.readme.novels.repository;
-
-import com.readme.novels.model.Novels;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface INovelsRepository extends JpaRepository<Novels, Long> {
-
-}
