--- src/main/java/com/readme/novels/responseObject/Response.java
@@ -1,15 +1,0 @@
-package com.readme.novels.responseObject;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class Response {
-
-    private Object data;
-
-}
