--- src/main/java/com/readme/novels/responseObject/ResponseNovelsPagination.java
@@ -1,30 +1,0 @@
-package com.readme.novels.responseObject;
-
-import java.util.List;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-import lombok.ToString;
-
-@Getter
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-@ToString
-public class ResponseNovelsPagination {
-    Object contents;
-    Object pagination;
-
-    @Getter
-    @Builder
-    @NoArgsConstructor
-    @AllArgsConstructor
-    public static class Pagination {
-        private int page;
-        private int size;
-        private long totalElements;
-        private int totalPage;
-    }
-}
