--- src/main/java/com/readme/utils/BaseTimeEntity.java
@@ -1,24 +1,0 @@
-package com.readme.utils;
-
-import java.time.LocalDateTime;
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-@MappedSuperclass
-@EntityListeners(AuditingEntityListener.class)
-@Getter
-public class BaseTimeEntity {
-
-    @CreatedDate
-    @Column(updatable = false)
-    private LocalDateTime createDate;
-    @LastModifiedDate
-    private LocalDateTime updateDate;
-
-
-}
