+++ src/test/java/com/readme/novel/NovelsApplicationTests.java
@@ -1,0 +1,13 @@
+package com.readme.novel;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class NovelsApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
