--- src/test/java/com/readme/novels/NovelsApplicationTests.java
@@ -1,13 +1,0 @@
-package com.readme.novels;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class NovelsApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
